# src/main/resources/design/product-swagger.yaml
openapi: 3.0.0
info:
  title: Product API
  description: API for managing products and their descriptions.
  version: 1.0.0

paths:
  /products:
    post:
      operationId: saveProduct
      summary: Create a new product with descriptions
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created successfully
        '400':
          description: Invalid input
    get:
      operationId: getAllProducts
      summary: Get all products with their descriptions
      tags:
        - Products
      responses:
        '200':
          description: A list of products with their descriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'

components:
  schemas:
    ProductRequest:
      type: object
      properties:
        productName:
          type: string
          description: Name of the product
        category:
          type: string
          description: Category of the product
        price:
          type: number
          format: double
          description: Price of the product
        stockQuantity:
          type: integer
          description: Quantity of the product in stock
        descriptions:
          type: array
          items:
            $ref: '#/components/schemas/ProductDescription'
    ProductResponse:
      type: object
      properties:
        productId:
          type: integer
          description: Unique identifier for the product
        productName:
          type: string
          description: Name of the product
        category:
          type: string
          description: Category of the product
        price:
          type: number
          format: double
          description: Price of the product
        stockQuantity:
          type: integer
          description: Quantity of the product in stock
        descriptions:
          type: array
          items:
            $ref: '#/components/schemas/ProductDescription'
    ProductDescription:
      type: object
      properties:
        languageCode:
          type: string
          description: Language of the description
        description:
          type: string
          description: Detailed description of the product